[mgsim] Produce more memory statistics.
git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3275 e97e5017-a994-416e-809d-76780e9f78db

+++ src/coma/simlink/linkprocs.cpp
@@ -15,8 +15,8 @@
 using namespace std;
 
 namespace MemSim{
-    unsigned int g_uMemoryAccessesL = 0;
-    unsigned int g_uMemoryAccessesS = 0;
+    uint64_t g_uMemoryAccessesL = 0;
+    uint64_t g_uMemoryAccessesS = 0;
 }
